+++ kotlin-compiler-extensions/main/kotlin/com/javiersc/kotlin/compiler/extensions/fir/coneKotlinType.kt
@@ -1,0 +1,12 @@
+package com.javiersc.kotlin.compiler.extensions.fir
+
+import org.jetbrains.kotlin.fir.symbols.impl.ConeClassLikeLookupTagImpl
+import org.jetbrains.kotlin.fir.types.ConeClassLikeType
+import org.jetbrains.kotlin.fir.types.ConeTypeProjection
+import org.jetbrains.kotlin.fir.types.impl.ConeClassLikeTypeImpl
+import org.jetbrains.kotlin.name.ClassId
+
+public fun ClassId.toConeType(vararg typeArguments: ConeTypeProjection): ConeClassLikeType {
+    val lookupTag = ConeClassLikeLookupTagImpl(this)
+    return ConeClassLikeTypeImpl(lookupTag, typeArguments, isNullable = false)
+}