+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- `ClassId.toConeType(): ConeClassLikeType`
+
 ### Changed
 
 - Kotlin version to `1.9.255-SNAPSHOT`